Add id to assembly fixes #10
Changing versions of the maven-assembly-plugin broke the build when
using newer versions. Versions after 2.4.1 (which is actually an ancient
version) require an assembly description to have an id. Set id to
'update-site-assembly' and set 'appendAssemblyId' to false. This uses a
correct id for the build, but stops maven from appending the id to the
finished zip. Just makes the file name unncesserily long and adds no
value.

I don't really know why this hasn't broken the build before, because we
never actually used a version older than 2.4.1. Will very likely be a
problem in builds that inherited from this project.